introduce dimension metadata
