+++ src/test/java/com/gooddata/md/NestedAttributeTest.java
@@ -3,27 +3,22 @@
  */
 package com.gooddata.md;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
+
 import org.codehaus.jackson.map.ObjectMapper;
 import org.hamcrest.Matchers;
 import org.testng.annotations.Test;
 
-import java.io.InputStream;
 import java.util.Collection;
 
-import static com.gooddata.JsonMatchers.serializesToJson;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-public class AttributeTest {
-
-    public static final String TITLE = "Person ID";
+public class NestedAttributeTest {
 
     @Test
-    public void shouldDeserialize() throws Exception {
-        final InputStream stream = getClass().getResourceAsStream("/md/attribute.json");
-        final Attribute attribute = new ObjectMapper().readValue(stream, Attribute.class);
+    public void testDeserialization() throws Exception {
+        final NestedAttribute attribute = new ObjectMapper().readValue(getClass().getResourceAsStream("/md/dimensionAttribute.json"), NestedAttribute.class);
         assertThat(attribute, is(notNullValue()));
 
         final Collection<DisplayForm> displayForms = attribute.getDisplayForms();
@@ -35,14 +30,14 @@ public class AttributeTest {
 
         assertThat(displayForm.getFormOf(), is("/gdc/md/PROJECT_ID/obj/DF_FORM_OF_ID"));
         assertThat(displayForm.getExpression(), is("[/gdc/md/PROJECT_ID/obj/DF_EXPRESSION_ID]"));
-        assertThat(displayForm.getLdmExpression(), is("[/gdc/md/PROJECT_ID/obj/DF_LDM_EXPRESSION_ID]"));
+        assertThat(displayForm.getLdmExpression(), isEmptyString());
 
         final DisplayForm defaultDisplayForm = attribute.getDefaultDisplayForm();
         assertThat(defaultDisplayForm, is(notNullValue()));
 
         assertThat(defaultDisplayForm.getFormOf(), is("/gdc/md/PROJECT_ID/obj/DF_FORM_OF_ID"));
         assertThat(defaultDisplayForm.getExpression(), is("[/gdc/md/PROJECT_ID/obj/DF_EXPRESSION_ID]"));
-        assertThat(defaultDisplayForm.getLdmExpression(), is("[/gdc/md/PROJECT_ID/obj/DF_LDM_EXPRESSION_ID]"));
+        assertThat(defaultDisplayForm.getLdmExpression(), isEmptyString());
 
         final Collection<Key> primaryKeys = attribute.getPrimaryKeys();
         assertThat(primaryKeys, is(Matchers.notNullValue()));
@@ -57,13 +52,4 @@ public class AttributeTest {
         assertThat(attribute.hasDimension(), is(true));
         assertThat(attribute.getDimensionLink(), is("/gdc/md/PROJECT_ID/obj/DIM_ID"));
     }
-
-    @Test
-    public void testSerialization() throws Exception {
-        final Attribute attribute = new Attribute(TITLE, new Key("/gdc/md/PROJECT_ID/obj/PK_ID", "col"),
-                new Key("/gdc/md/PROJECT_ID/obj/FK_ID", "col"));
-
-        assertThat(attribute, serializesToJson("/md/attribute-input.json"));
-    }
-
 }
