+++ src/test/resources/md/dimension-input.json
@@ -1,0 +1,13 @@
+{
+  "dimension" : {
+    "content" : {
+      "attributes" : []
+    },
+    "meta" : {
+      "title" : "Dimension",
+      "deprecated" : "0",
+      "locked": 0,
+      "unlisted": 0
+    }
+  }
+}
